refactor: move XML utilities from services to lib
RISDEV-7626
