+++ frontend/src/views/amending-law/articles/editor/AmendingLawArticleEditor.view.vue
@@ -14,10 +14,10 @@ import { useModHighlightClasses } from "@/composables/useZeitgrenzenHighlightCla
 import { useNormRenderHtml } from "@/composables/useNormRender"
 import { useNormXml } from "@/composables/useNormXml"
 import { getFrbrDisplayText } from "@/lib/frbr"
-import { getModEIds } from "@/services/ldmldeModService"
-import { getNodeByEid } from "@/services/ldmldeService"
+import { getModEIds } from "@/lib/ldmldeMod"
+import { getNodeByEid } from "@/lib/ldmlde"
 import { useGetNorm } from "@/services/normService"
-import { xmlNodeToString, xmlStringToDocument } from "@/services/xmlService"
+import { xmlNodeToString, xmlStringToDocument } from "@/lib/xml"
 import type { LawElementIdentifier } from "@/types/lawElementIdentifier"
 import { useDebounce } from "@vueuse/core"
 import type { Ref } from "vue"
